requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Django==4.2.11
django-bootstrap-v5==1.0.11
django-ckeditor==6.7.1
django-js-asset==2.2.0
pillow==10.2.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
